# contracts/openapi.json
{
  "components": {
    "schemas": {
      "AgentErrorCode": {
        "enum": [
          "MISSING_DATA",
          "STALE_INPUT",
          "PROVIDER_UNAVAILABLE",
          "BUDGET_EXCEEDED",
          "INVALID_AGENT_OUTPUT",
          "POLICY_DENIED",
          "DEADLINE_EXCEEDED"
        ],
        "title": "AgentErrorCode",
        "type": "string"
      },
      "AgentResult": {
        "additionalProperties": false,
        "properties": {
          "agent": {
            "enum": [
              "planning",
              "rm",
              "ie",
              "quality"
            ],
            "title": "Agent",
            "type": "string"
          },
          "data_quality": {
            "$ref": "#/components/schemas/DataQuality"
          },
          "error_code": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentErrorCode"
              },
              {
                "type": "null"
              }
            ]
          },
          "evidence_refs": {
            "items": {
              "$ref": "#/components/schemas/EvidenceRef"
            },
            "title": "Evidence Refs",
            "type": "array"
          },
          "execution_metadata": {
            "$ref": "#/components/schemas/ExecutionMetadata"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "title": "Findings",
            "type": "array"
          },
          "input_versions": {
            "additionalProperties": true,
            "title": "Input Versions",
            "type": "object"
          },
          "metrics": {
            "items": {
              "$ref": "#/components/schemas/Metric"
            },
            "title": "Metrics",
            "type": "array"
          },
          "recommended_actions": {
            "items": {
              "$ref": "#/components/schemas/RecommendedAction"
            },
            "title": "Recommended Actions",
            "type": "array"
          },
          "schema_version": {
            "const": "1.0",
            "title": "Schema Version",
            "type": "string"
          },
          "status": {
            "enum": [
              "SUCCEEDED",
              "DEGRADED",
              "FAILED"
            ],
            "title": "Status",
            "type": "string"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "summary_source": {
            "enum": [
              "deterministic",
              "model"
            ],
            "title": "Summary Source",
            "type": "string"
          },
          "task_id": {
            "format": "uuid",
            "title": "Task Id",
            "type": "string"
          },
          "warnings": {
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "schema_version",
          "task_id",
          "agent",
          "status",
          "summary",
          "summary_source",
          "findings",
          "metrics",
          "recommended_actions",
          "evidence_refs",
          "warnings",
          "input_versions",
          "data_quality",
          "execution_metadata"
        ],
        "title": "AgentResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LlmLabel": {
        "properties": {
          "is_fixture": {
            "title": "Is Fixture",
            "type": "boolean"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "provider": {
            "title": "Provider",
            "type": "string"
          }
        },
        "required": [
          "provider",
          "model",
          "is_fixture",
          "label"
        ],
        "title": "LlmLabel",
        "type": "object"
      },
...
      "OrderRef": {
        "properties": {
          "external_ref": {
            "title": "External Ref",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "external_ref"
        ],
        "title": "OrderRef",
        "type": "object"
      },
...
      "Page_RunSummary_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/RunSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[RunSummary]",
        "type": "object"
      },
...
      "RecommendationOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "generated_by": {
            "title": "Generated By",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "proposal_hash": {
            "title": "Proposal Hash",
            "type": "string"
          },
          "proposed_by_agent": {
            "title": "Proposed By Agent",
            "type": "string"
          },
          "rationale": {
            "title": "Rationale",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "superseded_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Superseded Reason"
          }
        },
        "required": [
          "id",
          "kind",
          "status",
          "generated_by",
          "proposed_by_agent",
          "rationale",
          "proposal_hash",
          "expires_at",
          "superseded_reason",
          "created_at"
        ],
        "title": "RecommendationOut",
        "type": "object"
      },
...
      "RunAccepted": {
        "properties": {
          "run_id": {
            "format": "uuid",
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "status"
        ],
        "title": "RunAccepted",
        "type": "object"
      },
      "RunDetail": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "deadline_at": {
            "format": "date-time",
            "title": "Deadline At",
            "type": "string"
          },
          "degraded_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Degraded Reason"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "llm": {
            "$ref": "#/components/schemas/LlmLabel"
          },
          "model_calls_limit": {
            "title": "Model Calls Limit",
            "type": "integer"
          },
          "model_calls_used": {
            "title": "Model Calls Used",
            "type": "integer"
          },
          "order": {
            "$ref": "#/components/schemas/OrderRef"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/RecommendationOut"
            },
            "title": "Recommendations",
            "type": "array"
          },
          "replan_count": {
            "title": "Replan Count",
            "type": "integer"
          },
          "report": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report"
          },
          "requested_by": {
            "$ref": "#/components/schemas/UserRef"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/AgentResult"
            },
            "title": "Results",
            "type": "array"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/TaskOut"
            },
            "title": "Tasks",
            "type": "array"
          },
          "tokens_used": {
            "title": "Tokens Used",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "order",
          "status",
          "requested_by",
          "llm",
          "degraded_reason",
          "error_code",
          "created_at",
          "completed_at",
          "model_calls_used",
          "model_calls_limit",
          "tokens_used",
          "replan_count",
          "started_at",
          "deadline_at",
          "tasks",
          "results",
          "recommendations",
          "report"
        ],
        "title": "RunDetail",
        "type": "object"
      },
      "RunEventOut": {
        "properties": {
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          }
        },
        "required": [
          "id",
          "event_type",
          "actor",
          "payload",
          "created_at"
        ],
        "title": "RunEventOut",
        "type": "object"
      },
      "RunSummary": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "degraded_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Degraded Reason"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "llm": {
            "$ref": "#/components/schemas/LlmLabel"
          },
          "order": {
            "$ref": "#/components/schemas/OrderRef"
          },
          "requested_by": {
            "$ref": "#/components/schemas/UserRef"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "id",
          "order",
          "status",
          "requested_by",
          "llm",
          "degraded_reason",
          "error_code",
          "created_at",
          "completed_at"
        ],
        "title": "RunSummary",
        "type": "object"
      },
...
      "UserRef": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "display_name"
        ],
        "title": "UserRef",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "LineSense AI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/orders/{order_id}/runs": {
      "get": {
        "operationId": "list_order_runs_api_v1_orders__order_id__runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "order_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Order Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_RunSummary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Order Runs",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/v1/runs/{run_id}": {
      "get": {
        "operationId": "get_run_api_v1_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}/cancel": {
      "post": {
        "operationId": "cancel_run_api_v1_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}/events": {
      "get": {
        "operationId": "list_run_events_api_v1_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_id",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "After Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunEventOut"
                  },
                  "title": "Response List Run Events Api V1 Runs  Run Id  Events Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Run Events",
        "tags": [
          "runs"
        ]
      }
    },
    "/api/v1/runs/{run_id}/retry": {
      "post": {
        "operationId": "retry_run_api_v1_runs__run_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 8,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Run",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}